+++ src/AreaAvatar.js
@@ -1,32 +1,16 @@
-import React, {useState} from 'react';
+import React from 'react';
 import {Avatar} from './Avatar';
-import {PlaceholderAvatar} from './PlaceholderAvatar';
 
-export function AreaAvatar({avatar, name, onClickAvatar}) {
-    const [isErrorLoadImg, setIsErrorLoadImg] = useState(false);
+export function AreaAvatar(props) {
+    let {avatar, onClickAvatar} = props;
 
-    let initials = name && name.trim().split(' ', 2).map(n => n[0]).join('').toUpperCase();
-
-    function onErrorLoad() {
-        setIsErrorLoadImg(true);
-    }
-
-    return(
+    return (
         <div 
             onClick={onClickAvatar} 
             style={{cursor: onClickAvatar ? 'pointer' : null}}
             class="avatar flex-center flex-column"
         >
-            {((avatar && !isErrorLoadImg) && 
-                <Avatar 
-                    onErrorLoad={onErrorLoad} 
-                    avatar={avatar} 
-                />
-            ) || 
-            (initials && 
-                <div class="avatar-initials flex-center">{initials}</div>
-            )
-            || <PlaceholderAvatar />}
+            <Avatar {...props} key={avatar} />
         </div>
     )
 }
