+++ src/Avatar.js
@@ -1,14 +1,33 @@
-import React from 'react';
+import React, {useState} from 'react';
+import {PlaceholderAvatar} from './PlaceholderAvatar';
 
-import './App.css';
+export function Avatar({avatar, name}) {
+    const [isErrorLoadImg, setIsErrorLoadImg] = useState(false);
 
-export function Avatar({avatar, onErrorLoad}) {
-    return (
-        <img 
-            class="avatar-image" 
-            src={avatar} 
-            alt=""
-            onError={onErrorLoad}
-        />
-    )
+    let initials = name && name.trim().split(' ', 2).map(n => n[0]).join('').toUpperCase();
+
+    function onErrorLoad() {
+        setIsErrorLoadImg(true);
+    }
+
+    if (avatar && !isErrorLoadImg) {
+        return (
+            <img 
+                class="avatar-image" 
+                src={avatar} 
+                alt=""
+                onError={onErrorLoad}
+            />
+        )
+    }
+    else if (initials) {
+        return (
+            <div class="avatar-initials flex-center">{initials}</div>
+        )
+    }
+    else {
+        return (
+            <PlaceholderAvatar />
+        )
+    }
 }
