+++ src/Card.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
-
-import {Avatar} from './Avatar';
+import {Avatar} from './Avatar'
 
 import './App.css';
